# api/src/openapi.yaml
openapi: 3.0.2
info:
  version: 0.1.0
  title: NFT Gallery API
  description: API serving the NFT Gallery Platform
...
security:
  - bearerAuth: []
servers:
  - url: '{protocol}://{host}:{port}'
    description: 'Local Mocks (e.g. http://localhost:4010) and Local Dev Server (e.g. http://localhost:3000/v1)'
    variables:
      protocol:
        enum:
          - http
          - https
        default: http
        description: Protocol of the URL.
      host:
        enum:
          - api-dev.domi.house
          - api.domi.house
          - localhost
        default: localhost
        description: Host of the URL
      port:
        enum:
          - ''
          - '4010'
          - '3000'
        default: '3000'
        description: Port of the URL
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Authentication token
paths:
...
  /v1/bid:
    post:
      summary: Create Bid
      operationId: createBid
      x-eov-operation-handler: BidsController
      x-security-scopes:
        - admin
        - consumer
      tags:
        - Bid
      requestBody:
        description: information that is needed for Bid
        required: true
        content:
          application/json:
            schema:
              title: Bid
              type: object
              description: Nft Gallery User Bid for an Asset
              x-autogen:
                openapi: true
                ownership: true
                mongo: true
                security:
                  create:
                    - admin
                    - consumer
                  update:
                    - admin
                    - consumer
                  delete:
                    - admin
                    - consumer
                  retrieve:
                    - admin
                    - consumer
                  retrieves:
                    - admin
                    - consumer
              properties:
                ownerId:
                  type: string
                assetId:
                  type: string
                bidderId:
                  type: string
                price:
                  type: number
                date:
                  type: string
              required:
                - assetId
                - bidderId
                - price
                - date
      responses:
        '200':
          description: Operation Executed
        default:
          description: A standard error object.
          content:
            application/json:
              schema:
                properties:
                  code:
                    type: string
                    x-faker: random.number
                  message:
                    type: string
                    example: Sample error message
                required:
                  - code
                x-tags:
                  - Common
  /v1/bids:
    parameters:
      - name: limit
        in: query
        description: How many items to return at one time (max 100)
        required: false
        schema:
          type: integer
          format: int32
      - name: skip
        in: query
        description: How many items should be skipped before returnig the required amount
        required: false
        schema:
          type: integer
          format: int32
    get:
      summary: Get all Bids
      operationId: retrieveBids
      x-eov-operation-handler: BidsController
      x-security-scopes:
        - admin
        - consumer
      tags:
        - Bid
      parameters:
        - name: ownerId
          in: query
          required: false
          description: The Fileter ownerId
          schema:
            type: string
            example: null
        - name: assetId
          in: query
          required: false
          description: The Fileter assetId
          schema:
            type: string
            example: null
        - name: bidderId
          in: query
          required: false
          description: The Fileter bidderId
          schema:
            type: string
            example: null
        - name: price
          in: query
          required: false
          description: The Fileter price
          schema:
            type: string
            example: 1.5
        - name: date
          in: query
          required: false
          description: The Fileter date
          schema:
            type: string
            example: null
      responses:
        '200':
          description: Operation Executed
        default:
          description: A standard error object.
          content:
            application/json:
              schema:
                properties:
                  code:
                    type: string
                    x-faker: random.number
                  message:
                    type: string
                    example: Sample error message
                required:
                  - code
                x-tags:
                  - Common
  '/v1/bids/{bidId}':
    parameters:
      - name: bidId
        in: path
        required: true
        description: The Id of the Bid
        schema:
          description: Id of the entity
          type: string
          format: not-empty-string
          example: 60758e80d59c357138d01af6
    get:
      summary: Get Bid with Id
      operationId: retrieveBid
      x-eov-operation-handler: BidsController
      x-security-scopes:
        - admin
        - consumer
      tags:
        - Bid
      parameters:
        - name: bidId
          in: path
          required: true
          description: The Id of the Bid
          schema:
            description: Id of the entity
            type: string
            format: not-empty-string
            example: 60758e80d59c357138d01af6
      responses:
        '200':
          description: Operation Executed
        default:
          description: A standard error object.
          content:
            application/json:
              schema:
                properties:
                  code:
                    type: string
                    x-faker: random.number
                  message:
                    type: string
                    example: Sample error message
                required:
                  - code
                x-tags:
                  - Common
    put:
      summary: Update Bid with Id
      operationId: updateBid
      x-eov-operation-handler: BidsController
      x-security-scopes:
        - admin
        - consumer
      tags:
        - Bid
      parameters:
        - name: bidId
          in: path
          required: true
          description: The Id of the Bid
          schema:
            description: Id of the entity
            type: string
            format: not-empty-string
            example: 60758e80d59c357138d01af6
      requestBody:
        description: information that is needed for Bid
        required: true
        content:
          application/json:
            schema:
              title: Update Bid with Id
              type: object
              description: An array of fields that we want to change name and value
              properties:
                ownerId:
                  type: string
                assetId:
                  type: string
                bidderId:
                  type: string
                price:
                  type: number
                date:
                  type: string
      responses:
        '200':
          description: Operation Executed
        default:
          description: A standard error object.
          content:
            application/json:
              schema:
                properties:
                  code:
                    type: string
                    x-faker: random.number
                  message:
                    type: string
                    example: Sample error message
                required:
                  - code
                x-tags:
                  - Common
    delete:
      summary: Delete Bid with Id
      operationId: deleteBid
      x-eov-operation-handler: BidsController
      x-security-scopes:
        - admin
        - consumer
      tags:
        - Bid
      parameters:
        - name: bidId
          in: path
          required: true
          description: The Id of the Bid
          schema:
            description: Id of the entity
            type: string
            format: not-empty-string
            example: 60758e80d59c357138d01af6
      responses:
        '200':
          description: Operation Executed
        default:
          description: A standard error object.
          content:
            application/json:
              schema:
                properties:
                  code:
                    type: string
                    x-faker: random.number
                  message:
                    type: string
                    example: Sample error message
                required:
                  - code
                x-tags:
                  - Common
...